% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nace_2_nace_11conc.R
\docType{data}
\name{nace_2_nace_11conc}
\alias{nace_2_nace_11conc}
\title{nace_2_nace_11conc.}
\format{
A data.table with 88 rows and 3 variables:
\describe{
\item{nace_2_2}{NACE v1 code at the two digit level}
\item{nace_1_2}{NACE v2 code at the two digit level}
\item{overlap_pro}{proportion of the codes that overalp between nace_2_2 and nace_1_2 at the 4 digit level}
}
}
\source{
\url{https://ec.europa.eu/eurostat/ramon/relations/index.cfm?TargetUrl=LST_REL&StrLanguageCode=EN&IntCurrentPage=11}
}
\usage{
nace_2_nace_11conc
}
\description{
A dataset containing mapping of NACE 2 to NACE 1.1 with and indication of the proportion of the codes that overlap
between the two digit code at lower levels of aggregation.
}
\keyword{datasets}
