% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uksic2003_uksic2007conc.R
\docType{data}
\name{uksic2003_uksic2007conc}
\alias{uksic2003_uksic2007conc}
\title{uksic2003_uksic2007conc.}
\format{
A data.table with 85 rows and 3 variables:
\describe{
\item{sic2007}{UK SIC 2007 code at the two digit level}
\item{sic2003}{UK SIC 2003 code at the two digit level}
\item{count}{proportion of the codes that overalp between nace_2_2 and nace_1_2 at the 4 digit level}
}
}
\source{
\url{https://www.ons.gov.uk/methodology/classificationsandstandards/ukstandardindustrialclassificationofeconomicactivities/uksic2007}
}
\usage{
uksic2003_uksic2007conc
}
\description{
A dataset containing mapping of UK SIC 2003 to UK SIC 2007 with and indication of the proportion of the codes that overlap
between the two digit code at lower levels of aggregation.
}
\keyword{datasets}
