% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussic_isic3conc.R
\docType{data}
\name{ussic_isic3conc}
\alias{ussic_isic3conc}
\title{ussic_isic3conc.}
\format{
A data.table with 93 rows and 3 variables:
\describe{
\item{isic_3_2}{ISIC V3.1 code at the two digit level}
\item{us_sic_2}{US SIC 2003 code at the two digit level}
\item{overlap btw isic3 & ussic}{proportion of the codes that overalp between isic_3_2 and us_sic_2 at the 4 digit level}
}
}
\source{
\url{https://unstats.un.org/unsd/classifications/Econ/ISIC.cshtml}
}
\usage{
ussic_isic3conc
}
\description{
A dataset containing mapping of US SIC 2007 to ISIC V3.1 with and indication of the proportion of the codes that overlap
between the two digit code at lower levels of aggregation.
}
\keyword{datasets}
